+++ astroquery/alma/core.py
@@ -19,6 +19,7 @@ from astroquery import log
 from astropy.utils.console import ProgressBar
 from astropy import units as u
 from astropy.time import Time
+from astropy.coordinates import SkyCoord
 
 from pyvo.dal.sia2 import SIA2_PARAMETERS_DESC, SIA2Service
 
@@ -32,7 +33,7 @@ from .tapsql import (_gen_pos_sql, _gen_str_sql, _gen_numeric_sql,
 from . import conf, auth_urls
 from astroquery.exceptions import CorruptDataWarning
 
-__all__ = {'AlmaClass', 'ALMA_BANDS'}
+__all__ = {'AlmaClass', 'ALMA_BANDS', 'get_enhanced_table'}
 
 __doctest_skip__ = ['AlmaClass.*']
 
@@ -207,6 +208,106 @@ def _gen_sql(payload):
     return sql + where
 
 
+def get_enhanced_table(result):
+    """
+    Returns an enhanced table with quantities instead of values and regions for footprints.
+    Note that this function is dependent on the ``astropy`` - affiliated ``regions`` package.
+    """
+    try:
+        import regions
+    except ImportError:
+        print(
+            "Could not import astropy-regions, which is a requirement for get_enhanced_table function in alma."
+            "Please refer to http://astropy-regions.readthedocs.io/en/latest/installation.html for how to install it.")
+
+    def _parse_stcs_string(input):
+        csys = 'icrs'
+
+        def _get_region(tokens):
+            if tokens[0] == 'polygon':
+                if csys == tokens[1].lower():
+                    tokens = tokens[2:]
+                else:
+                    tokens = tokens[1:]
+                points = SkyCoord(
+                    [(float(tokens[ii]), float(tokens[ii + 1])) * u.deg
+                     for ii in range(0, len(tokens), 2)], frame=csys)
+                return regions.PolygonSkyRegion(points)
+            elif tokens[0] == 'circle':
+                if csys == tokens[1].lower():
+                    tokens = tokens[2:]
+                else:
+                    tokens = tokens[1:]
+                return regions.CircleSkyRegion(
+                    SkyCoord(float(tokens[0]), float(tokens[1]), unit=u.deg),
+                    float(tokens[2]) * u.deg)
+            else:
+                raise ValueError("Unrecognized shape: " + tokens[0])
+        s_region = input.lower().strip()
+        if s_region.startswith('union'):
+            res = None
+            # skip the union operator
+            input_regions = s_region[s_region.index('(') + 1:s_region.rindex(
+                ')')].strip()
+            # omit the first char in the string to force it look for the second
+            # occurrence
+            last_pos = None
+            not_operation = False  # not operation - signals that the next substring is just the not operation
+            not_shape = False  # not shape - signals that the next substring is a not shape
+            for shape in re.finditer(r'not|polygon|circle', input_regions):
+                pos = shape.span()[0]
+                if last_pos is None:
+                    last_pos = pos
+                    continue  # this is the first elem
+                if shape.group() == 'not':
+                    not_operation = True
+                elif not_operation:
+                    not_shape = True
+                    not_operation = False
+                    last_pos = pos
+                    continue
+                if res is not None:
+                    next_shape = _get_region(
+                        input_regions[last_pos:pos - 1].strip(' ()').split())
+                    if not_shape:
+                        res = (res or next_shape) ^ next_shape
+                        not_shape = False
+                    else:
+                        res = res | next_shape
+                else:
+                    res = _get_region(
+                        input_regions[last_pos:pos - 1].strip().split())
+                last_pos = pos
+            if last_pos:
+                next_shape = _get_region(
+                    input_regions[last_pos:].strip(' ()').split())
+                res = res | next_shape
+            return res
+        elif 'not' in s_region:
+            # shape with "holes"
+            comps = s_region.split('not')
+            result = _get_region(comps[0].strip(' ()').split())
+            for comp in comps[1:]:
+                hole = _get_region(comp.strip(' ()').split())
+                result = (result or hole) ^ hole
+            return result
+        else:
+            return _get_region(s_region.split())
+    prep_table = result.to_qtable()
+    s_region_parser = None
+    for field in result.resultstable.fields:
+        if ('s_region' == field.ID) and \
+                ('obscore:Char.SpatialAxis.Coverage.Support.Area' == field.utype):
+            if 'adql:REGION' == field.xtype:
+                s_region_parser = _parse_stcs_string
+            # this is where to add other xtype parsers such as shape
+            break
+    if (s_region_parser):
+        for row in prep_table:
+            row['s_region'] = s_region_parser(row['s_region'])
+    return prep_table
+
+
 class AlmaAuth(BaseVOQuery, BaseQuery):
     """Authentication session information for passing credentials to an OIDC instance
 
@@ -376,7 +477,8 @@ class AlmaClass(QueryWithLogin):
         return self._tap_url
 
     def query_object_async(self, object_name, *, public=True,
-                           science=True, payload=None, **kwargs):
+                           science=True, payload=None, enhanced_results=False,
+                           **kwargs):
         """
         Query the archive for a source name.
 
@@ -390,6 +492,9 @@ class AlmaClass(QueryWithLogin):
         science : bool
             True to return only science datasets, False to return only
             calibration, None to return both
+        enhanced_results : bool
+            True to return a table with quantities instead of just values. It
+            also returns the footprint as `regions` objects.
         payload : dict
             Dictionary of additional keywords.  See `help`.
         """
@@ -398,10 +503,12 @@ class AlmaClass(QueryWithLogin):
         else:
             payload = {'source_name_resolver': object_name}
         return self.query_async(public=public, science=science,
-                                payload=payload, **kwargs)
+                                payload=payload, enhanced_results=enhanced_results,
+                                **kwargs)
 
     def query_region_async(self, coordinate, radius, *, public=True,
-                           science=True, payload=None, **kwargs):
+                           science=True, payload=None, enhanced_results=False,
+                           **kwargs):
         """
         Query the ALMA archive with a source name and radius
 
@@ -419,6 +526,9 @@ class AlmaClass(QueryWithLogin):
             calibration, None to return both
         payload : dict
             Dictionary of additional keywords.  See `help`.
+        enhanced_results : bool
+            True to return a table with quantities instead of just values. It
+            also returns the footprints as `regions` objects.
         """
         rad = radius
         if not isinstance(radius, u.Quantity):
@@ -433,11 +543,12 @@ class AlmaClass(QueryWithLogin):
             payload['ra_dec'] = ra_dec
 
         return self.query_async(public=public, science=science,
-                                payload=payload, **kwargs)
+                                payload=payload, enhanced_results=enhanced_results,
+                                **kwargs)
 
     def query_async(self, payload, *, public=True, science=True,
                     legacy_columns=False, get_query_payload=False,
-                    maxrec=None, **kwargs):
+                    maxrec=None, enhanced_results=False, **kwargs):
         """
         Perform a generic query with user-specified payload
 
@@ -458,6 +569,9 @@ class AlmaClass(QueryWithLogin):
             Flag to indicate whether to simply return the payload.
         maxrec : integer
             Cap on the amount of records returned.  Default is no limit.
+        enhanced_results : bool
+            True to return a table with quantities instead of just values. It
+            also returns the footprints as `regions` objects.
 
         Returns
         -------
@@ -492,7 +606,10 @@ class AlmaClass(QueryWithLogin):
         result = self.query_tap(query, maxrec=maxrec)
 
         if result is not None:
-            result = result.to_table()
+            if enhanced_results:
+                result = get_enhanced_table(result)
+            else:
+                result = result.to_table()
         else:
             # Should not happen
             raise RuntimeError('BUG: Unexpected result None')